Refactoring after review: 1) Removed jdbcTemplate usages, used NamedParameterJdbcTemplate instead in all dao classes; 2) Added final modifier for fields on entities and dtos where possible; 3) Removed @autowired annotations where possible and used dependency injection through the constructor into finalized fields; 4) Removed unnecessary import and unused class members;

+++ src/main/java/org/nomarch/movieland/dao/jdbc/mapper/GenreRowMapper.java
@@ -10,7 +10,7 @@ public class GenreRowMapper implements RowMapper<Genre> {
     @Override
     public Genre mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         return Genre.builder()
-                .id(resultSet.getInt("id"))
+                .id(resultSet.getLong("id"))
                 .name(resultSet.getString("name"))
                 .build();
     }
